test(synthesizer): evmSetup with USDT storage layout
