+++ packages/frontend/synthesizer/constants/storage-layouts/USDC_IMP.json
@@ -1,0 +1,181 @@
+{
+  "storageLayout": {
+    "storage": [
+      {
+        "astId": 1,
+        "contract": "FiatTokenV1",
+        "label": "_owner",
+        "offset": 0,
+        "slot": "0",
+        "type": "address"
+      },
+      {
+        "astId": 2,
+        "contract": "FiatTokenV1",
+        "label": "_paused",
+        "offset": 0,
+        "slot": "1",
+        "type": "bool"
+      },
+      {
+        "astId": 3,
+        "contract": "FiatTokenV1",
+        "label": "_rescuer",
+        "offset": 0,
+        "slot": "2",
+        "type": "address"
+      },
+      {
+        "astId": 4,
+        "contract": "FiatTokenV1",
+        "label": "_name",
+        "offset": 0,
+        "slot": "3",
+        "type": "string"
+      },
+      {
+        "astId": 5,
+        "contract": "FiatTokenV1",
+        "label": "_symbol",
+        "offset": 0,
+        "slot": "4",
+        "type": "string"
+      },
+      {
+        "astId": 6,
+        "contract": "FiatTokenV1",
+        "label": "_decimals",
+        "offset": 0,
+        "slot": "5",
+        "type": "uint8"
+      },
+      {
+        "astId": 7,
+        "contract": "FiatTokenV1",
+        "label": "_masterMinter",
+        "offset": 0,
+        "slot": "6",
+        "type": "address"
+      },
+      {
+        "astId": 8,
+        "contract": "FiatTokenV1",
+        "label": "balances",
+        "offset": 0,
+        "slot": "7",
+        "type": "mapping(address => uint256)"
+      },
+      {
+        "astId": 9,
+        "contract": "FiatTokenV1",
+        "label": "allowed",
+        "offset": 0,
+        "slot": "8",
+        "type": "mapping(address => mapping(address => uint256))"
+      },
+      {
+        "astId": 10,
+        "contract": "FiatTokenV1",
+        "label": "totalSupply_",
+        "offset": 0,
+        "slot": "9",
+        "type": "uint256"
+      },
+      {
+        "astId": 11,
+        "contract": "FiatTokenV1",
+        "label": "minters",
+        "offset": 0,
+        "slot": "10",
+        "type": "mapping(address => bool)"
+      },
+      {
+        "astId": 12,
+        "contract": "FiatTokenV1",
+        "label": "minterAllowed",
+        "offset": 0,
+        "slot": "11",
+        "type": "mapping(address => uint256)"
+      },
+      {
+        "astId": 13,
+        "contract": "FiatTokenV2",
+        "label": "_balanceAndBlacklistStates",
+        "offset": 0,
+        "slot": "12",
+        "type": "mapping(address => uint256)"
+      },
+      {
+        "astId": 14,
+        "contract": "FiatTokenV2",
+        "label": "_authorizations",
+        "offset": 0,
+        "slot": "13",
+        "type": "mapping(address => mapping(address => bool))"
+      },
+      {
+        "astId": 15,
+        "contract": "FiatTokenV2_1",
+        "label": "_nonces",
+        "offset": 0,
+        "slot": "14",
+        "type": "mapping(address => uint256)"
+      }
+    ],
+    "types": {
+      "address": {
+        "encoding": "inplace",
+        "label": "address",
+        "numberOfBytes": "20"
+      },
+      "bool": {
+        "encoding": "inplace",
+        "label": "bool",
+        "numberOfBytes": "1"
+      },
+      "string": {
+        "encoding": "bytes",
+        "label": "string",
+        "numberOfBytes": "32"
+      },
+      "uint8": {
+        "encoding": "inplace",
+        "label": "uint8",
+        "numberOfBytes": "1"
+      },
+      "uint256": {
+        "encoding": "inplace",
+        "label": "uint256",
+        "numberOfBytes": "32"
+      },
+      "mapping(address => uint256)": {
+        "encoding": "mapping",
+        "key": "address",
+        "label": "mapping(address => uint256)",
+        "numberOfBytes": "32",
+        "value": "uint256"
+      },
+      "mapping(address => bool)": {
+        "encoding": "mapping",
+        "key": "address",
+        "label": "mapping(address => bool)",
+        "numberOfBytes": "32",
+        "value": "bool"
+      },
+      "mapping(address => mapping(address => uint256))": {
+        "encoding": "mapping",
+        "key": "address",
+        "label": "mapping(address => mapping(address => uint256))",
+        "numberOfBytes": "32",
+        "value": "mapping(address => uint256)"
+      },
+      "mapping(address => mapping(address => bool))": {
+        "encoding": "mapping",
+        "key": "address",
+        "label": "mapping(address => mapping(address => bool))",
+        "numberOfBytes": "32",
+        "value": "mapping(address => bool)"
+      }
+    }
+  }
+}
